Add: Home 부분 데이터 바인딩 완료
추후 작업으로 데이터가 존재하지 않을경우 NoneFeed 보이게 할 예정입니다.
