+++ src/components/home/HomeContainer/index.style.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const ProfileSection = styled.main`
+  position: fixed;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 20px 16px 0;
+  height: calc(100% - 108px);
+  overflow-y: scroll;
+  min-width: 390px;
+  width: 100%;
+  min-width: 100%;
+  height: 100%;
+  margin-top: 20px;
+`;
