--- src/components/home/ProfileContainer/index.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-
-function ProfileContainer() {
-  return <div>index</div>;
-}
-export default ProfileContainer;
