--- src/components/home/ProfileContainer/index.style.js
@@ -1,1 +1,0 @@
-import styled from "styled-components";
