+++ src/components/profile/ProfileContainer/index.jsx
@@ -3,7 +3,6 @@ import { useDispatch, useSelector } from "react-redux";
 import { UserInfoContainer } from "./index.style";
 import { useHistory, useParams } from "react-router-dom";
 import { getUserMyProfile } from "../../../actions/userActions";
-
 import ProfileCard from "../ProfileCard";
 
 function ProfileContainer() {
